requirements.txt
# Core Dependencies
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
gspread==6.0.0

# AI/ML
anthropic==0.18.1
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.20
chromadb==0.4.22
sentence-transformers==2.3.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
schedule==1.2.0
retry==0.9.2

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Monitoring & Logging
loguru==0.7.2
